move files to sensible locations
