+++ src/lexer/LexError.cpp
