+++ src/parser/ParseError.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <stdexcept>
-#include "../core/ASTNode.hpp"
+#include "../ast/ASTNode.hpp"
 
 namespace zenith {
