+++ tools/count_non_comment_lines.ps1
